-  add ternary operator on switch
