+++ ChatApp/Presentation/View/ViewController/MessageTableViewCell.swift
@@ -22,14 +22,12 @@ class MessageTableViewCell: UITableViewCell {
     
     override func awakeFromNib() {
         super.awakeFromNib()
-        // Initialization code
-      //  self.reuseIdentifier = "MessageTableViewCell"
+      
     }
 
     override func setSelected(_ selected: Bool, animated: Bool) {
         super.setSelected(selected, animated: animated)
 
-        // Configure the view for the selected state
     }
 
 }
