--- ChatApp/Presentation/View/sendCustomView/SendMessageView/sendButton/SendButtonConstants.swift
@@ -1,15 +1,0 @@
-////
-////  SendButtonConstants.swift
-////  ChatApp
-////
-////  Created by MacBoobPro on 02.05.23.
-////
-//
-//import Foundation
-//import UIKit
-//
-//extension SendButtonView {
-//    enum Constants {
-//        static let image = UIImage(named: "send_button_image")
-//    }
-//}
